Add:gameover,case,bug fiks

+++ AlkonostXNA/AlkonostXNA/AlkonostDataStructure/Interfaces/IAttackable.cs
@@ -7,7 +7,7 @@ namespace AlkonostXNAGame.AlkonostDataStructure.Interfaces
 {
     public interface IAttackable
     {
-        float Hit();
+        int Hit();
         float Defend();
     }
 }
